libs/py-common/pyproject.toml
[project]
name = "agent-platform-common"
version = "0.1.0"
description = "Shared library for Agent Platform microservices"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.8",
  "pydantic-settings>=2.4",
  "structlog>=24.0",
  "opentelemetry-api>=1.25",
  "opentelemetry-sdk>=1.25",
  "opentelemetry-exporter-otlp>=1.25",
  "opentelemetry-instrumentation-fastapi>=0.46b0",
  "opentelemetry-instrumentation-httpx>=0.46b0",
  "opentelemetry-instrumentation-redis>=0.46b0",
  "opentelemetry-instrumentation-psycopg2>=0.46b0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_platform_common"]
